requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
black==22.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cytoolz==0.11.2
discord.py==1.7.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
isort==5.10.1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.19.4
pycryptodome==3.14.1
pyrsistent==0.18.1
python-dateutil==2.8.2
requests==2.27.1
rlp==2.0.1
six==1.16.0
strictyaml==1.6.1
tomli==2.0.1
toolz==0.11.2
typing_extensions==4.1.1
urllib3==1.26.8
varint==1.0.2
web3==5.28.0
websockets==9.1
yarl==1.7.2
